Refactor controller to service
